debounced buttons + quick tick thru timer 0

+++ code/src/time_face.h
@@ -4,7 +4,7 @@
 struct TimeFace
 {
   bool countSeconds = false;
-  void loop(uint8_t event);
+  void loop(uint16_t event);
   void drawTime(bool forceRedraw);
 };
